feat: improve MonthlyAggregationGraph
